<feat> Add Parser struct w/ first pass impl
This commit adds the parser struct with the ability to perform a first pass of the token list created from the scanner. The first pass constructs the symbol table which is a table of all the labels in the assembly code and their respective addresses.  It also implements a small fraction of the instruction parsing logic in the second pass. The second pass will use a dispatch table that maps opcodes to parsing functions for each instruction.